Refactoring, front end imagen request updated to Axios to facilitate bearer token, added debugging on imagen backend for testing
